+++ tests/MaxMind/Test/MinFraud/UtilTest.php
@@ -158,6 +158,16 @@ class UtilTest extends TestCase
                     ],
                 ],
             ],
+            [
+                'name' => 'Equivalent domain',
+                'input' => ['email' => ['address' => 'foo@googlemail.com']],
+                'expected' => [
+                    'email' => [
+                        'address' => md5('foo@gmail.com'),
+                        'domain' => 'gmail.com',
+                    ],
+                ],
+            ],
         ];
 
         if (\function_exists('idn_to_ascii')