+++ CHANGELOG.md
@@ -5,6 +5,8 @@ CHANGELOG
 ------------------
 
 * Added `pxp_financial` and `trustpay` to the payment processor validation.
+* Equivalent domain names are now normalized when `hashEmail` is used.
+  For example, `googlemail.com` will become `gmail.com`.
 
 2.0.0 (2023-12-04)
 ------------------